Creating the Bond Tile Component
Created the Bond Tile component to display a list of bonds that a user may subscribe to. Currently not implemented in any part of the app and with no styling but most of the relevant JSX produced.

This is a reusable component to be called in a list of available bonds.