Separated the approve and added documentation
A new function has been created to request approval for the bond. This function should be called first when subscribing to get approval, it returns the address of the bond that gets passed to the Dai.approve() function to allow the bond to make the transfer. After that, the subscriberToBond() function can be called.